Extract factorial function to shared utility
